Moved image data to util
