+++ talkgenerator/schema/presentation_schema.py
@@ -9,7 +9,7 @@ from multiprocessing.pool import ThreadPool
 import random
 from typing import List, Collection, Callable, Dict, Union, Optional
 
-from schema.image_data import ImageData
+from util.image_data import ImageData
 from schema.slide_topic_generators import SlideSeedGenerator
 from talkgenerator.schema.slide_generator_data import (
     _filter_generated_elements,
