+++ talkgenerator/slide/powerpoint_slide_creator.py
@@ -1,6 +1,5 @@
 import os
 import sys
-import traceback
 import logging
 from functools import lru_cache
 from typing import List
@@ -8,7 +7,7 @@ from typing import List
 from lxml.etree import XMLSyntaxError
 from pptx import Presentation
 
-from schema.image_data import ImageData
+from util.image_data import ImageData
 from talkgenerator.util import os_util
 
 # Location of powerpoint template
