+++ talkgenerator/util/image_data.py
